chore: rename to Navbar for its types and interfaces
