+++ frontend/src/constants/themes.ts
@@ -1,3 +1,3 @@
-import { type Theme } from "@/types/layout";
+import { type Theme } from '@/types/navbar';
 
 export const DEFAULT_THEME: Theme = 'light';
