+++ frontend/src/types/navbar.ts
